Make goto-definition work correctly when the symbol is in a macro arg.
Consider:

  SOME_MACRO(foobar());

You place the cursor on 'foobar' and go to definition.

Before this change, you would end up jumping into the definition of SOME_MACRO, to the first place where its argument is substituted. This is not useful. If you wanted to go to SOME_MACRO, you'd have clicked on it, not on foobar.

After this change, you will jump to the definition of foobar().